# Add mode to ignore Rust generic type parameters