… rust_expected_test.go