Depending on the situation it may be too expensive to fully demangle a function name and all of its generic type parameters. In particular Rust very aggressively inlines and tries to make abstractions zero-cost, which can result in extreme nesting of generic type-parameters.

An example function name from a real life workload demonstrates this, for which a benchmark has been added. The difference of enabling/disabling the newly introduced mode is rather dramatic (even if expected since the mode intentionally does less).

```
$ benchstat with-generic-params.txt without-generic-params.txt
name     old time/op  new time/op  delta
Rust-10   109µs ± 0%     2µs ± 0%  -97.98%  (p=0.008 n=5+5)
```