@ianlancetaylor 2023-04-25
Thanks.  I did this somewhat differently.  I reused the existing `NoTemplateParams`, making it work for the Rust demangler as well.  I just dropped everything inside the <>.  I applied it to "dyn" entries as well.

Let me know if you see any problems.  Thanks.