@brancz 2023-04-26
I don't think I feel strongly, the only thing is that now we just get empty `<>`, while previously we would at least see the number of type parameters taken.

Perf wise it seems identical:

```
$ benchstat without-types.txt without-types-no-template.txt
name     old time/op  new time/op  delta
Rust-10  2.19µs ± 0%  2.23µs ± 0%  +1.67%  (p=0.008 n=5+5)
```