@ianlancetaylor 2023-04-26
Yes, it seemed to me that `<>` was clearer than `<T>`.  After all, there really could be a type `T`.  And if you don't care about the type parameters, why do you care exactly how many there are?

But I'm not a Rust programmer, so let me know if I'm mistaken.